feat: feed logs to loki using loki logback appender, add banner

+++ src/main/java/com/bookwhile/author/service/AuthorService.java
@@ -49,6 +49,7 @@ public class AuthorService {
         AuthorEntity authorEntity = authorRepository.findById(id)
             .orElseThrow(() -> new BookWhileException(AUTHOR_NOT_FOUND));
 
+
         authorDtoMapper.updateAuthorEntity(authorEntity, authorRequestDto);
 
         authorRepository.save(authorEntity);
@@ -56,8 +57,9 @@ public class AuthorService {
 
     public void deleteAuthor(UUID id) {
 
-        authorRepository.findById(id)
-            .orElseThrow(() -> new BookWhileException(AUTHOR_NOT_FOUND));
+        if (!authorRepository.existsById(id)) {
+            throw new BookWhileException(AUTHOR_NOT_FOUND);
+        }
 
         authorRepository.deleteById(id);
     }
